refactor verse suggesting
