--- src/suggesetor/getSuggestionFromVod.ts
@@ -1,3 +1,0 @@
-import { VerseSuggesting } from '../VerseSuggesting'
-
-export const getSuggestionFromVod = async (): Promise<VerseSuggesting[]> => {}
